# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mystics ERP API — Sales & CRM, Project Management, Inventory & Warehouse Management
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login and get JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /crm-invoices:
    get:
      operationId: getCrmInvoices
      tags: [crm-invoices]
      summary: List CRM invoices
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: clientPoId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CrmInvoice"
    post:
      operationId: createCrmInvoice
      tags: [crm-invoices]
      summary: Create a client invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrmInvoiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmInvoice"
  /crm-invoices/{id}/mark-paid:
    post:
      operationId: markCrmInvoicePaid
      tags: [crm-invoices]
      summary: Mark invoice as paid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkPaidInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmInvoice"
...
  /escalations:
    get:
      operationId: getEscalations
      tags: [escalations]
      summary: List escalations
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: projectId
          in: query
          schema:
            type: integer
        - name: module
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Escalations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Escalation"
    post:
      operationId: createEscalation
      tags: [escalations]
      summary: Raise an escalation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EscalationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Escalation"
  /escalations/{id}/resolve:
    post:
      operationId: resolveEscalation
      tags: [escalations]
      summary: Resolve an escalation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveEscalationInput"
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Escalation"

  # ── PROJECTS ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        orgId:
          type: integer

    # ── DASHBOARD ─────────────────────────────────────────────────────────────
...
    CrmInvoice:
      type: object
      required: [id, clientPoId, type, amount, paymentStatus, createdAt]
      properties:
        id:
          type: integer
        clientPoId:
          type: integer
        projectId:
          type: ["integer", "null"]
        type:
          type: string
        amount:
          type: number
        taxDetails:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        paymentStatus:
          type: string
        createdAt:
          type: string
    CrmInvoiceInput:
      type: object
      required: [clientPoId, type, amount]
      properties:
        clientPoId:
          type: integer
        projectId:
          type: integer
        type:
          type: string
        amount:
          type: number
        taxDetails:
          type: string
        dueDate:
          type: string
    MarkPaidInput:
      type: object
      required: [paidAmount, paidDate]
      properties:
        paidAmount:
          type: number
        paidDate:
          type: string

    # ── TASKS ─────────────────────────────────────────────────────────────────
...
    Escalation:
      type: object
      required: [id, reason, severity, status, createdAt]
      properties:
        id:
          type: integer
        sourceEntityType:
          type: ["string", "null"]
        sourceEntityId:
          type: ["integer", "null"]
        projectId:
          type: ["integer", "null"]
        module:
          type: ["string", "null"]
        raisedBy:
          type: ["integer", "null"]
        raisedByName:
          type: ["string", "null"]
        reason:
          type: string
        severity:
          type: string
        assignedTo:
          type: ["integer", "null"]
        assignedToName:
          type: ["string", "null"]
        status:
          type: string
        resolvedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    EscalationInput:
      type: object
      required: [reason, severity]
      properties:
        sourceEntityType:
          type: string
        sourceEntityId:
          type: integer
        projectId:
          type: integer
        module:
          type: string
        reason:
          type: string
        severity:
          type: string
        assignedTo:
          type: integer
    ResolveEscalationInput:
      type: object
      required: [resolution]
      properties:
        resolution:
          type: string

    # ── PROJECTS ──────────────────────────────────────────────────────────────
...